implemented chatservice

+++ backend/src/main/java/com/nicolas/chatapp/dto/request/GroupChatRequestDTO.java
@@ -2,5 +2,5 @@ package com.nicolas.chatapp.dto.request;
 
 import java.util.List;
 
-public record GroupChatRequestDTO(List<Integer> userIds, String chatName) {
+public record GroupChatRequestDTO(List<Long> userIds, String chatName) {
 }
